starting to work on the page attributes
this method uses the "any" operator to search through the string of text
where all of the attributes we're interested in get stored and returns a
boolean value, "TRUE" or "FALSE". It's nifty because it easily allows us
to create dummy variables for the attributes we're interested in.  We
would need to modify this to be more inclusive of the different
possibilities, but I think it will be a good approach!